Reorganizing test structure.
